requirements.txt
PAM==0.4.2
Pillow==2.3.0
PyBluez==0.18
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
ipython==1.2.1
lockfile==0.8
lxml==3.3.3
matplotlib==1.3.1
numpy==1.8.2
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
pexpect==3.1
piston-mini-client==0.7.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pygobject==3.12.0
pyparsing==2.0.1
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5ubuntu1
python-dateutil==1.5
python-debian==0.1.21-nmu2ubuntu2
pytz==2012c
pyxdg==0.25
reportlab==3.0
requests==2.2.1
scikit-learn==0.16.1
scipy==0.13.3
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.5.2
software-center-aptd-plugins==0.0.0
system-service==0.1.6
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==13.1.2
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
